requirements.txt
# RAISA — Dependencias Python
# Instalar con: pip install -r requirements.txt

# Discord bot framework
discord.py>=2.3.0

# SQLite async driver
aiosqlite>=0.19.0

# Variables de entorno
python-dotenv>=1.0.0

# Procesamiento de imágenes (compresión de avatares)
Pillow>=10.0.0

# HTTP async (verificación de URLs de imagen)
aiohttp>=3.9.0